Add files via upload
A project work with E-novia Italy new technology. This technology is represented by a device with haptic sensors that can mimic the touch. A new solution was crafted using platform thinking, technolog epiphany and design thinking. This file is the presentation that was prepared for a delegation from E-novia and the course professors.